requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
jinja2>=3.1.4
aiofiles>=24.1.0
pytest>=8.3.0
pytest-asyncio>=0.23.0
